LayoutTests/fast/device-orientation/device-motion-request-permission-granted-expected.txt

CONSOLE MESSAGE: No device motion events will be fired, reason: Permission to use the API was not yet requested.
Received device orientation & motion access request for top level origin "", with frame origin "".
Basic testing for DeviceMotionEvent.requestPermission().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result is "granted"
PASS result is "granted"
PASS successfullyParsed is true

TEST COMPLETE
